+++ tools/targets/Program.cs
@@ -95,7 +95,7 @@ internal sealed class Program
                 var versionFromReleaseNotes = File.ReadLines(releaseNotesFile, Encoding.UTF8)
                     .First(line => line.StartsWith("## ", StringComparison.Ordinal)).Substring(3).Trim();
                 Console.WriteLine($"Read version '{versionFromReleaseNotes}' from release notes");
-                var tagName = Environment.GetEnvironmentVariable("APPVEYOR_REPO_TAG_NAME");
+                var tagName = Environment.GetEnvironmentVariable("GITHUB_REF_NAME");
                 if (versionFromReleaseNotes != tagName)
                 {
                     Console.WriteLine($"Release notes version does not match tag name '{tagName}'. Disambiguating.");
@@ -105,8 +105,8 @@ internal sealed class Program
                     }
 
                     version = versionFromReleaseNotes +
-                        "+Build." + (Environment.GetEnvironmentVariable("APPVEYOR_BUILD_NUMBER") ?? "adhoc") +
-                        "-Sha." + (Environment.GetEnvironmentVariable("APPVEYOR_REPO_COMMIT") ?? "adhoc");
+                        "+Build." + (Environment.GetEnvironmentVariable("GITHUB_RUN_NUMBER") ?? "adhoc") +
+                        "-Sha." + (Environment.GetEnvironmentVariable("GITHUB_SHA") ?? "adhoc");
                 }
                 else
                 {